remove renderprops on mark as primary flow
